LayoutTests/imported/w3c/web-platform-tests/workers/interfaces/WorkerGlobalScope/location/redirect-sharedworker-expected.txt

CONSOLE MESSAGE: line 8: ReferenceError: Can't find variable: SharedWorker

Harness Error (FAIL), message = ReferenceError: Can't find variable: SharedWorker
